+++ packages/frontend/src/components/Template/ListItem.vue
@@ -5,7 +5,7 @@
       {{ translation.data.title }}
     </v-list-item-title>
     <v-list-item-subtitle>
-      <chip-review-status :value="translation.status" />
+      <chip-status :value="translation.status" />
       <s-tooltip v-for="tr in props.template.translations" :key="tr.id">
         <template #activator="{ props: tooltipProps }">
           <chip-language :value="tr.language" v-bind="tooltipProps" />