+++ packages/nuxt-base-layer/src/components/DesignListItem.vue
@@ -1,7 +1,7 @@
 <template>
   <v-list-item :to="`/designs/${item.id}/pdfdesigner/`" :title="title" lines="two" :data-testid="`design-${item.id}`">
     <v-list-item-subtitle>
-      <chip-review-status v-if="item.status" :value="item.status" />
+      <chip-status v-if="item.status" :value="item.status" />
       <chip-language v-if="item.language" :value="item.language" />
       <chip-tag v-for="tag in props.item.tags || []" :key="tag" :value="tag" />
     </v-list-item-subtitle>